backend/requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.12.4,<4.0
# For environment variable management
python-dotenv>=0.19,<1.0